Unflake E2E spec due to missing taxonomy label (#845)
This fixes the failure in our E2E spec:

```
ActionView::Template::Error:
  ERROR in SERVER PRERENDERING
  Encountered error:

  Error evaluating server bundle. Check your webpack configuration.
  ===============================================================
  Caught error:
  Error: Expected root.taxonomies.294.label to be a string, got null

``` 

Taxonomy labels are maintained in 

https://github.com/antiwork/gumroad/blob/e3725cac86f4620cc12d72ff3a99c28e55e05868/app/presenters/discover/taxonomy_presenter.rb#L4-L8

Taxonomies are JSON-ified via

https://github.com/antiwork/gumroad/blob/e3725cac86f4620cc12d72ff3a99c28e55e05868/app/presenters/discover/taxonomy_presenter.rb#L327-L343

Thus, new taxonomies with labels outside of
`TaxonomyPresenter::TAXONOMY_LABELS` will product a JSON with `null`
label, failing the TypeScript type check, resulting in the test failure.

So, this PR fixes the failure by using an existing taxonomy rather than
creating a new one.

I have also considered filtering out taxonomies with `null` labels (and
logging an error to Rails & Bugsnag), but decided to keep this as-is so
we can't / won't be able to ignore the big red failure in test & dev
envs.