RSpec feature spec -> system spec (#892)
This migrates our browser-driven specs from `:feature` to `:system`.

This made it possible to remove DatabaseCleaner in favor of
[`use_transactional_fixtures`](https://rspec.info/features/7-1/rspec-rails/Transactions/)
-- seems to have made E2E specs 10% faster.

Interestingly, this also helped uncovered some legit wrong E2E tests
that were not failing before.

Some other misc changes I made:

- Updated `selenium-webdriver` to the latest version
- Removed `webdrivers` as a dependency (thanks to the
`selenium-webdriver` update)
- `JSErrorReporter` is now working again. But I had to disable it / make
it opt-in, as we have quite a number of failures from this.
- Added debugging logs for some common flaky specs