Access failed specs screenshot (#1137)
Part of: https://github.com/antiwork/gumroad/issues/1127

Some capybara browser tests that were failing flakily and inconsistently
on Github Actions, but I could not manage to reproduce locally at all.
What was going on? It would be super helpful to have access to the
capybara-screenshots generated on the github actions run.

:warning: This is super important to understand/debug how a spec failed
on the CI :warning:

---------

Co-authored-by: Ershad Kunnakkadan <ershad92@gmail.com>