Fix flaky post-purchase flash message in specs (#1246)
Ref: #1127 

We get a bunch of errors like this triggered from various specs:

<img width="2012" height="361" alt="Screenshot 2025-09-15 at 00 18 21"
src="https://github.com/user-attachments/assets/6bdce0aa-4824-4ccd-bac0-4884e7f18c05"
/>

The issue seems to be that the check for text on the purchase success
page occasionally happens after the alert has disappeared (maybe
Capybara waits for a random long-running network request or something?
seemed easier to make it work regardless of wait time rather than debug
the wait itself)

The fix is to check for the alert regardless of visibility, which works
because we don't reset the text when we hide it, but that requires
different logic for gifts and preorders where we show the receipt page
instead.