Improve logic fetching max_refund_period_in_days for PurchaseRefundPolicy records (#1254)
Part of #1047, follow-up on #1050

The one-time script uses AI (OpenAI `gpt-4o-mini`) to determine the
value for `max_refund_period_in_days`). With the number of records we
have in production, the cost of running this script could be
significant.

This PR adds new logic that tries to determine the value for
`max_refund_period_in_days` in 2 ways before using AI (which should
reduce the use of AI significantly):
* refund policy on the product, if the refund policy `title` hasn't
changed
* the first `PurchaseRefundPolicy` record that has the same title, and a
value set for `max_refund_period_in_days`