Chore/remove analytics from server components (#1617)
Follow up of a PR review in
https://github.com/antiwork/gumroad/pull/1609#discussion_r2419657071

This pull request removes the `Analytics` component from the list of SSR
components powered by React on Rails, since we no longer use React on
Rails but Inertia to render it.

It also:
- Renames the inertia template from `index` to `Index` to match
Inertia's conventions and avoid confusion with how `index` files are
used
- Moves the Analytics template from `app/components/server-components`
to `app/components` for consistency

### AI Disclosure

No AI was used at all.